commands/requirements.txt
requests
pyyaml
python-decouple